backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6

# IBM Granite AI Dependencies
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=4.21.0

# Optional: For better performance
# bitsandbytes>=0.41.0  # Uncomment for GPU acceleration
